Final contribution to the project. A new project will be made using C++ instead of C
The whole point of the project was to further my knowledge of C and to challenge myself to make a user-friendly version of Solitaire in the terminal. Now that I have acquired the knowledge I wish to acquire in C, I now just want to focus on C++ because I find it easier to program in.